flask_app/requirements.txt
PyJWT~=2.1.0
Flask~=1.1.2
Werkzeug~=1.0.1
SQLAlchemy==1.3.18
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
PyMySQL==1.0.2